+++ src/test/java/ch/psi/wica/infrastructure/stream/WicaStreamConfigurationDecoderTest.java
@@ -113,13 +113,13 @@ class WicaStreamConfigurationDecoderTest
       String testString = "{\"channels\":" +
             "[{\"name\":\"XHIPA:TIME\"},{\"name\":\"EVEX:STR:2\"},{\"name\":\"EWBRI:IST:2\"}," +
             "{\"name\":\"MXC1:IST:2\"},{\"name\":\"MYC2:IST:2\"},{\"name\":\"MHC4:IST:2\"},{\"name\":\"MHC6:IST:2\"}," +
-            "{\"name\":\"MHC1:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}}," +
-            "{\"name\":\"MYC2:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}}," +
-            "{\"name\":\"MBC1:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}}," +
-            "{\"name\":\"MRI12:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":0.3}}," +
-            "{\"name\":\"MII7:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":0.3}}," +
-            "{\"name\":\"MRI13:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":8}}," +
-            "{\"name\":\"MRI14:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":8}}," +
+            "{\"name\":\"MHC1:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}}," +
+            "{\"name\":\"MYC2:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}}," +
+            "{\"name\":\"MBC1:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}}," +
+            "{\"name\":\"MRI12:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":0.3}}," +
+            "{\"name\":\"MII7:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":0.3}}," +
+            "{\"name\":\"MRI13:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":8}}," +
+            "{\"name\":\"MRI14:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":8}}," +
             "{\"name\":\"CIPHMO:SOL:1\"},{\"name\":\"CRPHFT:SOL:1\"},{\"name\":\"MXF1:IST:2\"},{\"name\":\"MRFEIN:IST:2\"},{\"name\":\"MRFAUS:IST:2\"}," +
             "{\"name\":\"MII7:ILOG:2\"},{\"name\":\"MRI2:ILOG:2\"},{\"name\":\"MRI13:ILOG:2\"},{\"name\":\"MRI14:ILOG:2\"},{\"name\":\"CI1V:IST:2\"}," +
             "{\"name\":\"CI3V:IST:2\"},{\"name\":\"CI2V:IST:2\"},{\"name\":\"CI4V:IST:2\"},{\"name\":\"CR1V:IST:2\"},{\"name\":\"CR2V:IST:2\"}," +
@@ -128,7 +128,7 @@ class WicaStreamConfigurationDecoderTest
             "{\"name\":\"EECI:IST:2\"},{\"name\":\"CIREV:FIST:2\"},{\"name\":\"CRREV:FIST:2\"},{\"name\":\"AIHS:IST:2\"}," +
             "{\"name\":\"HS:IST:2\"},{\"name\":\"M3ALT:IST:2\"},{\"name\":\"GLS:LEISTUNG_AKTUELL\"}," +
             "{\"name\":\"ZSLP:TOTSAVEFAST\"}]," +
-            "\"props\":{\"heartbeat\":15000,\"monflux\":100,\"pollflux\":1000,\"daqmode\":\"monitor\",\"pollratio\":1,\"prec\":6,\"fields\":\"val;sevr\"}}";
+            "\"props\":{\"heartbeat\":15000,\"monflux\":100,\"pollflux\":1000,\"daqmode\":\"monitor\",\"pollint\":1000,\"prec\":6,\"fields\":\"val;sevr\"}}";
 
       final WicaStream wicaStream = decoder.decode( testString );
       final WicaStreamProperties wicaStreamProperties = wicaStream.getWicaStreamProperties();