+++ src/test/java/ch/psi/wica/infrastructure/stream/WicaStreamSerializerTest.java
@@ -231,42 +231,6 @@ class WicaStreamSerializerTest
       assertThat( props.getOptionalFilterDeadband().get(),                           is( 11.0 ) );
    }
 
-   @Test
-   void testDeserializeWicaStreamProperties_DeprecatedInterfaceTest_HeartBeatAliasWorks()
-   {
-      final String inputString = "{\"heartbeat\":\"15\"}";
-      final WicaStreamProperties props = WicaStreamSerializer.readFromJson( inputString, WicaStreamProperties.class );
-
-      assertThat( props.getOptionalHeartbeatFluxIntervalInMillis().isPresent(), is( true ) );
-      assertThat( props.getOptionalHeartbeatFluxIntervalInMillis().get(),       is( 15 ) );
-   }
-
-   @Test
-   void testDeserializeWicaStreamProperties_DeprecatedInterfaceTest_PollRatioIsIgnoredWhenPollingIntervalPresent()
-   {
-      final String inputString = "{\"pollflux\":\"15\",\"pollratio\":4 }";
-      final WicaStreamProperties props = WicaStreamSerializer.readFromJson( inputString, WicaStreamProperties.class );
-
-      assertThat( props.getOptionalPolledValueFluxIntervalInMillis().isPresent(),    is( true ) );
-      assertThat( props.getOptionalPollingIntervalInMillis().isPresent(),            is( true ) );
-      assertThat( props.getOptionalPolledValueFluxIntervalInMillis().get(),          is( 15 ) );
-      assertThat( props.getOptionalPollingIntervalInMillis().get(),                  is( 60 ) );
-   }
-
-   @Test
-   void testDeserializeWicaStreamProperties_DeprecatedInterfaceTest_PollRatioIsUsedWhenPollingIntervalAbsent()
-   {
-      final String inputString = "{\"pollflux\":\"15\",\"pollint\":5,\"pollratio\":4 }";
-      final WicaStreamProperties props = WicaStreamSerializer.readFromJson( inputString, WicaStreamProperties.class );
-
-      assertThat( props.getOptionalPolledValueFluxIntervalInMillis().isPresent(),    is( true ) );
-      assertThat( props.getOptionalPollingIntervalInMillis().isPresent(),            is( true ) );
-
-      assertThat( props.getOptionalPolledValueFluxIntervalInMillis().get(),          is( 15 ) );
-      assertThat( props.getOptionalPollingIntervalInMillis().get(),                  is( 5 ) );
-   }
-
-
    /****************************************************************************************
     * 3.0 WicaChannelProperties - Serialization Tests
     ****************************************************************************************/
@@ -381,7 +345,7 @@ class WicaStreamSerializerTest
    @Test
    void testDeserializeWicaStream_HIPA_Example()
    {
-      final String inputString = "   {\"channels\":[{\"name\":\"XHIPA:TIME\"},{\"name\":\"EVEX:STR:2\"},{\"name\":\"EWBRI:IST:2\"},{\"name\":\"MXC1:IST:2\"},{\"name\":\"MYC2:IST:2\"},{\"name\":\"MHC4:IST:2\"},{\"name\":\"MHC6:IST:2\"},{\"name\":\"MHC1:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}},{\"name\":\"MYC2:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}},{\"name\":\"MBC1:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}},{\"name\":\"MRI12:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":0.3}},{\"name\":\"MII7:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":0.3}},{\"name\":\"MRI13:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":8}},{\"name\":\"MRI14:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollratio\":30,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":8}},{\"name\":\"CIPHMO:SOL:1\"},{\"name\":\"CRPHFT:SOL:1\"},{\"name\":\"MXF1:IST:2\"},{\"name\":\"MRFEIN:IST:2\"},{\"name\":\"MRFAUS:IST:2\"},{\"name\":\"MII7:ILOG:2\"},{\"name\":\"MRI2:ILOG:2\"},{\"name\":\"MRI13:ILOG:2\"},{\"name\":\"MRI14:ILOG:2\"},{\"name\":\"CI1V:IST:2\"},{\"name\":\"CI3V:IST:2\"},{\"name\":\"CI2V:IST:2\"},{\"name\":\"CI4V:IST:2\"},{\"name\":\"CR1V:IST:2\"},{\"name\":\"CR2V:IST:2\"},{\"name\":\"CR3V:IST:2\"},{\"name\":\"CR4V:IST:2\"},{\"name\":\"CR5V:IST:2\"},{\"name\":\"UCNQ:BEAMREQ:STATUS\"},{\"name\":\"UCNQ:BEAMREQ:COUNTDOWN\"},{\"name\":\"EICV:IST:2\"},{\"name\":\"EICI:IST:2\"},{\"name\":\"EECV:IST:2\"},{\"name\":\"EECI:IST:2\"},{\"name\":\"CIREV:FIST:2\"},{\"name\":\"CRREV:FIST:2\"},{\"name\":\"AIHS:IST:2\"},{\"name\":\"HS:IST:2\"},{\"name\":\"M3ALT:IST:2\"},{\"name\":\"GLS:LEISTUNG_AKTUELL\"},{\"name\":\"ZSLP:TOTSAVEFAST\"}],\"props\":{\"heartbeat\":15000,\"monflux\":100,\"pollflux\":1000,\"daqmode\":\"monitor\",\"pollratio\":1,\"prec\":6,\"fields\":\"val;sevr\"}}";
+      final String inputString = "   {\"channels\":[{\"name\":\"XHIPA:TIME\"},{\"name\":\"EVEX:STR:2\"},{\"name\":\"EWBRI:IST:2\"},{\"name\":\"MXC1:IST:2\"},{\"name\":\"MYC2:IST:2\"},{\"name\":\"MHC4:IST:2\"},{\"name\":\"MHC6:IST:2\"},{\"name\":\"MHC1:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}},{\"name\":\"MYC2:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}},{\"name\":\"MBC1:IST:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":10}},{\"name\":\"MRI12:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":0.3}},{\"name\":\"MII7:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":0.3}},{\"name\":\"MRI13:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":8}},{\"name\":\"MRI14:ILOG:2##2\",\"props\":{\"daqmode\":\"poll\",\"pollint\":30000,\"fields\":\"val;ts\",\"filter\":\"changes\",\"deadband\":8}},{\"name\":\"CIPHMO:SOL:1\"},{\"name\":\"CRPHFT:SOL:1\"},{\"name\":\"MXF1:IST:2\"},{\"name\":\"MRFEIN:IST:2\"},{\"name\":\"MRFAUS:IST:2\"},{\"name\":\"MII7:ILOG:2\"},{\"name\":\"MRI2:ILOG:2\"},{\"name\":\"MRI13:ILOG:2\"},{\"name\":\"MRI14:ILOG:2\"},{\"name\":\"CI1V:IST:2\"},{\"name\":\"CI3V:IST:2\"},{\"name\":\"CI2V:IST:2\"},{\"name\":\"CI4V:IST:2\"},{\"name\":\"CR1V:IST:2\"},{\"name\":\"CR2V:IST:2\"},{\"name\":\"CR3V:IST:2\"},{\"name\":\"CR4V:IST:2\"},{\"name\":\"CR5V:IST:2\"},{\"name\":\"UCNQ:BEAMREQ:STATUS\"},{\"name\":\"UCNQ:BEAMREQ:COUNTDOWN\"},{\"name\":\"EICV:IST:2\"},{\"name\":\"EICI:IST:2\"},{\"name\":\"EECV:IST:2\"},{\"name\":\"EECI:IST:2\"},{\"name\":\"CIREV:FIST:2\"},{\"name\":\"CRREV:FIST:2\"},{\"name\":\"AIHS:IST:2\"},{\"name\":\"HS:IST:2\"},{\"name\":\"M3ALT:IST:2\"},{\"name\":\"GLS:LEISTUNG_AKTUELL\"},{\"name\":\"ZSLP:TOTSAVEFAST\"}],\"props\":{\"hbflux\":15000,\"monflux\":100,\"pollflux\":1000,\"daqmode\":\"monitor\",\"pollint\":30000,\"prec\":6,\"fields\":\"val;sevr\"}}";
       final WicaStream stream = WicaStreamSerializer.readFromJson( inputString, WicaStream.class );
       assertThat( stream.getWicaStreamProperties().getFieldsOfInterest(), is( "val;sevr") );
    }